app/requirements.txt
connexion[flask,swagger-ui]==3.*
gunicorn==21.*
Flask>=2.3,<3
Werkzeug<3
uvicorn[standard]
google-cloud-firestore==2.16.0
google-auth==2.33.0
google-cloud-logging==3.11.3